added conversation feature (which required some light to heavy refactoring) - pending code cleanup and protocol to conversation conversion

+++ src/main/java/jokrey/utilities/network/link2peer/node/message_headers/CustomExpirationHeader.java
@@ -8,7 +8,7 @@ import jokrey.utilities.network.link2peer.P2Link;
 public class CustomExpirationHeader extends MinimalHeader {
     private final short expiresAfter;
     private final long createdAtCtm;
-    public CustomExpirationHeader(P2Link sender, int type, short expiresAfter, boolean requestReceipt) {
+    public CustomExpirationHeader(P2Link sender, short type, short expiresAfter, boolean requestReceipt) {
         super(sender, type, requestReceipt);
         this.expiresAfter = expiresAfter;
         createdAtCtm = System.currentTimeMillis();
